Hora mostrada em horas

+++ src/pages/comercial/Produtos.tsx
@@ -697,7 +697,7 @@ export default function Produtos() {
           {tenantSettings.energy_cost_kwh === 0 && tenantSettings.labor_cost_hour === 0 && (
             <p className="text-[11px] text-muted-foreground">⚠ Configure custos de energia/mão de obra em Configurações → Empresa</p>
           )}
-          {!costBreakdown.hasMachineRate && parseInt(estTime) > 0 && (
+          {!costBreakdown.hasMachineRate && parseFloat(estTime) > 0 && (
             <p className="text-[11px] text-muted-foreground">⚠ Para calcular depreciação, preencha custo de aquisição e vida útil da impressora.</p>
           )}
         </div>
